+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed upgrade bug from pre-2.1.0 that could prevent correct binding to video and audio managers
 - Fixed editor error when sync audio profiles with no unity source defined in profile template
+- OptionsUI component can automatically detect parent video player / audio manager
 
 ## [2.1.6] - 09-23-23